+++ .secrets.baseline
@@ -3,7 +3,7 @@
     "files": "go.sum|^.secrets.baseline$",
     "lines": null
   },
-  "generated_at": "2025-03-12T13:08:42Z",
+  "generated_at": "2025-03-12T13:12:58Z",
   "plugins_used": [
     {
       "name": "AWSKeyDetector"
@@ -150,7 +150,7 @@
         "hashed_secret": "39f69c278f46165447f30d10acf54277aaa3d5fc",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 93,
+        "line_number": 94,
         "type": "Secret Keyword",
         "verified_result": null
       },
@@ -158,7 +158,7 @@
         "hashed_secret": "2ace62c1befa19e3ea37dd52be9f6d508c5163e6",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 252,
+        "line_number": 274,
         "type": "Secret Keyword",
         "verified_result": null
       }
@@ -177,7 +177,7 @@
         "hashed_secret": "39f69c278f46165447f30d10acf54277aaa3d5fc",
         "is_secret": false,
         "is_verified": false,
-        "line_number": 86,
+        "line_number": 79,
         "type": "Secret Keyword",
         "verified_result": null
       }
