+++ pkg/mounter/mounter.go
@@ -1,14 +1,24 @@
 package mounter
 
 import (
+	"context"
+	"encoding/json"
 	"errors"
+	"fmt"
+	"io"
+	"net"
+	"net/http"
 	"os"
+	"strings"
 	"syscall"
+	"time"
 
 	"github.com/IBM/ibm-object-csi-driver/pkg/constants"
 	mounterUtils "github.com/IBM/ibm-object-csi-driver/pkg/mounter/utils"
 	"k8s.io/klog/v2"
 )
+
+var mountWorker bool = true
 
 type Mounter interface {
 	Mount(source string, target string) error
@@ -105,4 +115,59 @@ var mkdirAllFunc = os.MkdirAll
 // Function that wraps os.MkdirAll
 var mkdirAll = func(path string, perm os.FileMode) error {
 	return mkdirAllFunc(path, perm)
+}
+
+func createMountHelperContainerRequest(payload string, url string) (string, error) {
+	timeout := 3 * time.Minute
+	defaultSocketPath := "/tmp/mysocket.sock"
+
+	// Get socket path
+	// socketPath := os.Getenv("SOCKET_PATH")
+	socketPath := "/var/lib/ibmshare.sock"
+	if socketPath == "" {
+		socketPath = defaultSocketPath
+	}
+	// Create a custom dialer function for Unix socket connection
+	dialer := func(ctx context.Context, network, addr string) (net.Conn, error) {
+		return net.Dial("unix", socketPath)
+	}
+
+	// Create an HTTP client with the Unix socket transport
+	client := &http.Client{
+		Transport: &http.Transport{
+			DialContext: dialer,
+		},
+		Timeout: timeout,
+	}
+
+	// Create POST request
+	req, err := http.NewRequest("POST", url, strings.NewReader(payload))
+	if err != nil {
+		return "", err
+	}
+	req.Header.Set("Content-Type", "application/json")
+	response, err := client.Do(req)
+	if err != nil {
+		return "", err
+	}
+	defer response.Body.Close()
+	body, err := io.ReadAll(response.Body)
+	if err != nil {
+		return "", err
+	}
+
+	// Unmarshal json response
+	var responseBody struct {
+		MountExitCode   string `json:"MountExitCode"`
+		ExitDescription string `json:"Description"`
+	}
+	err = json.Unmarshal(body, &responseBody)
+	if err != nil {
+		return "", err
+	}
+
+	if response.StatusCode != http.StatusOK {
+		return responseBody.ExitDescription, fmt.Errorf("response from mount-helper-container -> Exit Status Code: %s ,ResponseCode: %v", responseBody.MountExitCode, response.StatusCode)
+	}
+	return "", nil
 }
